+++ app-apple/Package/Sources/AppLibrary/Views/Paywall/CustomProductView.swift
@@ -8,6 +8,10 @@ import StoreKit
 import SwiftUI
 
 struct CustomProductView: View {
+
+    @Environment(\.isEnabled)
+    private var isEnabled
+
     let style: PaywallProductViewStyle
 
     @ObservedObject
@@ -23,12 +27,100 @@ struct CustomProductView: View {
     let onError: (Error) -> Void
 
     var body: some View {
+        contentView
+    }
+}
+
+private extension CustomProductView {
+    var contentView: some View {
+#if os(tvOS)
+        Button(action: purchase) {
+            VStack(alignment: .leading) {
+                Text(verbatim: product.localizedTitle)
+                    .themeTrailingValue(product.localizedPrice)
+                    .font(withDescription ? .headline : .footnote)
+
+                if withDescription {
+                    Text(verbatim: product.localizedDescription)
+                        .font(.subheadline)
+                        .foregroundStyle(.secondary)
+                }
+            }
+        }
+#else
         HStack {
-            Text(verbatim: product.localizedTitle)
+            VStack(alignment: .leading) {
+                Text(verbatim: product.localizedTitle)
+                    .font(isPrimary ? .title2 : withFooter ? .headline : nil)
+                    .fontWeight(isPrimary ? .bold : nil)
+
+                if withDescription {
+                    Text(verbatim: product.localizedDescription)
+                        .font(.subheadline)
+                        .foregroundStyle(.secondary)
+                }
+            }
             Spacer()
             Button(action: purchase) {
-                Text(verbatim: product.localizedPrice)
+                Text(product.localizedPrice)
+                    .font(.headline)
+                    .fontWeight(.bold)
+                    .padding(.horizontal, 15)
+                    .padding(.vertical, 5)
+                    .background(.quinary)
+                    .clipShape(.capsule)
+                    .foregroundStyle(isEnabled ? Color.accentColor : .gray)
+                    .cursor(.hand)
             }
+            .buttonStyle(.borderless)
+        }
+        .padding(withPadding ? 10 : .zero)
+#endif
+    }
+}
+
+private extension CustomProductView {
+    var isPurchasing: Bool {
+        purchasingIdentifier != nil
+    }
+
+    var isPrimary: Bool {
+        switch style {
+        case .donation:
+            false
+        case .paywall(let primary):
+            primary
+        }
+    }
+
+    var withDescription: Bool {
+        switch style {
+        case .donation:
+            false
+        case .paywall(let primary):
+            primary
+        }
+    }
+
+    var withFooter: Bool {
+        switch style {
+        case .donation:
+            false
+        case .paywall(let primary):
+#if os(tvOS)
+            primary
+#else
+            true // disclosing features
+#endif
+        }
+    }
+
+    var withPadding: Bool {
+        switch style {
+        case .donation:
+            true
+        case .paywall:
+            false
         }
     }
 }
@@ -47,5 +139,18 @@ private extension CustomProductView {
                 onError(error)
             }
         }
+    }
+}
+
+#Preview {
+    List {
+        CustomProductView(
+            style: .paywall(primary: true),
+            iapManager: .forPreviews,
+            product: AppProduct.Complete.OneTime.lifetime.asFakeIAP,
+            purchasingIdentifier: .constant(nil),
+            onComplete: { _, _ in },
+            onError: { _ in }
+        )
     }
 }
