+++ app-apple/Package/Sources/AppLibrary/Views/Paywall/PaywallView+Fixed.swift
@@ -63,7 +63,7 @@ private extension PaywallFixedView {
                 .font(.title2)
                 .padding(.bottom, 1)
             ForEach(model.completePurchasable, id: \.productIdentifier) { iap in
-                NewPaywallProductView(
+                PaywallProductView(
                     iapManager: iapManager,
                     style: .paywall(primary: true),
                     product: iap,
@@ -97,7 +97,7 @@ private extension PaywallFixedView {
                     .padding(.bottom, 1)
             }
             ForEach(model.individualPurchasable, id: \.productIdentifier) { iap in
-                NewPaywallProductView(
+                PaywallProductView(
                     iapManager: iapManager,
                     style: .paywall(primary: !showsComplete),
                     product: iap,
