+++ app-apple/Package/Sources/AppLibrary/Views/Paywall/PaywallView+Scrollable.swift
@@ -44,7 +44,7 @@ private extension PaywallScrollableView {
     var completeProductsView: some View {
         Group {
             ForEach(model.completePurchasable, id: \.productIdentifier) {
-                NewPaywallProductView(
+                PaywallProductView(
                     iapManager: iapManager,
                     style: .paywall(primary: true),
                     product: $0,
@@ -73,7 +73,7 @@ private extension PaywallScrollableView {
 
     var individualProductsView: some View {
         ForEach(model.individualPurchasable, id: \.productIdentifier) {
-            NewPaywallProductView(
+            PaywallProductView(
                 iapManager: iapManager,
                 style: .paywall(primary: false),
                 product: $0,
