+++ app-apple/Package/Sources/AppLibrary/Views/Paywall/PaywallView.swift
@@ -4,306 +4,59 @@
 
 import CommonLibrary
 import CommonUtils
-import StoreKit
 import SwiftUI
 
-#if !os(tvOS)
+struct PaywallView: View, SizeClassProviding {
 
-struct PaywallView: View {
+    @Environment(\.horizontalSizeClass)
+    var hsClass
 
-    @EnvironmentObject
-    private var theme: Theme
-
-    @EnvironmentObject
-    private var iapManager: IAPManager
+    @Environment(\.verticalSizeClass)
+    var vsClass
 
     @Binding
     var isPresented: Bool
 
+    @ObservedObject
+    var iapManager: IAPManager
+
     let requiredFeatures: Set<AppFeature>
 
-    // nil = essentials
-    let suggestedProducts: Set<AppProduct>?
+    @ObservedObject
+    var model: PaywallCoordinator.Model
 
-    @State
-    private var isFetchingProducts = true
+    @ObservedObject
+    var errorHandler: ErrorHandler
 
-    @State
-    private var products: [InAppProduct] = []
+    let onComplete: (String, InAppPurchaseResult) -> Void
 
-    @State
-    private var completeProducts: [InAppProduct] = []
-
-    @State
-    private var purchasingIdentifier: String?
-
-    @State
-    private var isPurchasePendingConfirmation = false
-
-    @StateObject
-    private var errorHandler: ErrorHandler = .default()
+    let onError: (Error) -> Void
 
     var body: some View {
-        contentView
-            .themeProgress(if: isFetchingProducts)
-            .toolbar(content: toolbarContent)
-            .alert(
-                Strings.Global.Actions.purchase,
-                isPresented: $isPurchasePendingConfirmation,
-                actions: pendingActions,
-                message: pendingMessage
+#if os(tvOS)
+        // TODO: #1511, use isBigDevice to also use fixed layout on macOS and iPad?
+//        if isBigDevice {
+            PaywallFixedView(
+                isPresented: $isPresented,
+                iapManager: iapManager,
+                requiredFeatures: requiredFeatures,
+                model: model,
+                errorHandler: errorHandler,
+                onComplete: onComplete,
+                onError: onError
             )
-            .task(id: suggestedProducts) {
-                await fetchAvailableProducts()
-            }
-            .withErrorHandler(errorHandler)
-    }
-}
-
-private extension PaywallView {
-    var title: String {
-        Strings.Global.Actions.purchase
-    }
-
-    var contentView: some View {
-        Form {
-            requiredFeaturesView
-            productsView
-            completeProductsView
-            restoreView
-            linksView
-        }
-        .themeForm()
-        .disabled(purchasingIdentifier != nil)
-    }
-
-    var requiredFeaturesView: some View {
-        requiredFeatures
-            .nilIfEmpty
-            .map { features in
-                FeatureListView(
-                    style: .list,
-                    header: Strings.Views.Paywall.Sections.RequiredFeatures.header,
-                    features: features.sorted(),
-                    content: {
-                        featureView(for: $0)
-                            .fontWeight(theme.relevantWeight)
-                    }
-                )
-            }
-    }
-
-    var productsView: some View {
-        products
-            .nilIfEmpty
-            .map { products in
-                ForEach(products, id: \.productIdentifier) {
-                    PaywallProductView(
-                        iapManager: iapManager,
-                        style: .paywall(),
-                        product: $0,
-                        highlightedFeatures: requiredFeatures,
-                        purchasingIdentifier: $purchasingIdentifier,
-                        onComplete: onComplete,
-                        onError: onError
-                    )
-                }
-                .themeSection(
-                    header: Strings.Views.Paywall.Sections.Products.header,
-                    footer: Strings.Views.Paywall.Sections.Products.footer
-                )
-            }
-    }
-
-    var completeProductsView: some View {
-        completeProducts
-            .nilIfEmpty
-            .map { products in
-                Group {
-                    ForEach(products, id: \.productIdentifier) {
-                        PaywallProductView(
-                            iapManager: iapManager,
-                            style: .paywall(),
-                            product: $0,
-                            withIncludedFeatures: false,
-                            highlightedFeatures: requiredFeatures,
-                            purchasingIdentifier: $purchasingIdentifier,
-                            onComplete: onComplete,
-                            onError: onError
-                        )
-                    }
-                    VStack(alignment: .leading) {
-                        ForEach(AppFeature.allCases.sorted()) {
-                            IncludedFeatureRow(
-                                feature: $0,
-                                isHighlighted: requiredFeatures.contains($0)
-                            )
-                            .font(.subheadline)
-                        }
-                    }
-                }
-                .themeSection(
-                    header: Strings.Views.Paywall.Sections.FullProducts.header,
-                    footer: [
-                        Strings.Views.Paywall.Sections.FullProducts.footer,
-                        Strings.Views.Paywall.Sections.Products.footer
-                    ].joined(separator: " ")
-                )
-            }
-    }
-
-    var linksView: some View {
-        Section {
-            Link(Strings.Unlocalized.eula, destination: Constants.shared.websites.eula)
-            Link(Strings.Views.Settings.Links.Rows.privacyPolicy, destination: Constants.shared.websites.privacyPolicy)
-        }
-    }
-
-    func featureView(for feature: AppFeature) -> some View {
-        Text(feature.localizedDescription)
-            .scrollableOnTV()
-    }
-
-    var restoreView: some View {
-        RestorePurchasesButton(errorHandler: errorHandler)
-            .themeContainerWithSingleEntry(
-                header: Strings.Views.Paywall.Sections.Restore.header,
-                footer: Strings.Views.Paywall.Sections.Restore.footer,
-                isAction: true
+//        } else {
+#else
+            PaywallScrollableView(
+                isPresented: $isPresented,
+                iapManager: iapManager,
+                requiredFeatures: requiredFeatures,
+                model: model,
+                errorHandler: errorHandler,
+                onComplete: onComplete,
+                onError: onError
             )
-    }
-
-    func toolbarContent() -> some ToolbarContent {
-        ToolbarItem(placement: .cancellationAction) {
-            Button {
-                isPresented = false
-            } label: {
-                ThemeCloseLabel()
-            }
-        }
-    }
-
-    func pendingActions() -> some View {
-        Button(Strings.Global.Nouns.ok) {
-            isPresented = false
-        }
-    }
-
-    func pendingMessage() -> some View {
-        Text(Strings.Views.Paywall.Alerts.Pending.message)
-    }
-}
-
-// MARK: -
-
-private extension PaywallView {
-    var essentialFeatures: [AppFeature] {
-        AppProduct.Essentials.iOS_macOS.features
-    }
-
-    func fetchAvailableProducts() async {
-        isFetchingProducts = true
-        defer {
-            isFetchingProducts = false
-        }
-        do {
-            let rawProducts = suggestedProducts ?? iapManager.suggestedProducts()
-            guard !rawProducts.isEmpty else {
-                throw AppError.emptyProducts
-            }
-            let rawCompleteProducts = rawProducts.filter(\.isComplete)
-
-            let allProducts = try await iapManager.purchasableProducts(for: rawProducts
-                .sorted {
-                    $0.productRank < $1.productRank
-                }
-            )
-            var products: [InAppProduct] = []
-            var completeProducts: [InAppProduct] = []
-            allProducts.forEach {
-                guard let raw = AppProduct(rawValue: $0.productIdentifier) else {
-                    return
-                }
-                if rawCompleteProducts.contains(raw) {
-                    completeProducts.append($0)
-                } else {
-                    products.append($0)
-                }
-            }
-
-            pp_log_g(.App.iap, .info, "Suggested products: \(products)")
-            guard !products.isEmpty || !completeProducts.isEmpty else {
-                throw AppError.emptyProducts
-            }
-
-            self.products = products
-            self.completeProducts = completeProducts
-        } catch {
-            pp_log_g(.App.iap, .error, "Unable to load purchasable products: \(error)")
-            onError(error, dismissing: true)
-        }
-    }
-
-    func onComplete(_ productIdentifier: String, result: InAppPurchaseResult) {
-        switch result {
-        case .done:
-            Task {
-                await iapManager.reloadReceipt()
-            }
-            isPresented = false
-
-        case .pending:
-            isPurchasePendingConfirmation = true
-
-        case .cancelled:
-            break
-
-        case .notFound:
-            fatalError("Product not found: \(productIdentifier)")
-        }
-    }
-
-    func onError(_ error: Error) {
-        onError(error, dismissing: false)
-    }
-
-    func onError(_ error: Error, dismissing: Bool) {
-        errorHandler.handle(error, title: Strings.Global.Actions.purchase) {
-            if dismissing {
-                isPresented = false
-            }
-        }
-    }
-}
-
-private extension AppProduct {
-    var productRank: Int {
-        switch self {
-        case .Essentials.iOS_macOS:
-            return .min
-        case .Essentials.iOS:
-            return 1
-        case .Essentials.macOS:
-            return 2
-        case .Complete.Recurring.yearly:
-            return 3
-        case .Complete.Recurring.monthly:
-            return 4
-        default:
-            return .max
-        }
-    }
-}
-
-// MARK: - Previews
-
-#Preview {
-    PaywallView(
-        isPresented: .constant(true),
-        requiredFeatures: [.appleTV],
-        suggestedProducts: [.Features.appleTV, .Complete.OneTime.lifetime]
-    )
-    .withMockEnvironment()
-}
-
+//        }
 #endif
+    }
+}
