+++ app-apple/Package/Sources/AppLibraryMain/Views/Profile/ProfileStorageSection.swift
@@ -59,12 +59,7 @@ private extension ProfileStorageSection {
                     Text(Strings.Unlocalized.iCloud)
                     PurchaseRequiredView(
                         requiring: sharingRequirements,
-                        reason: $paywallReason,
-                        suggesting: { // TODO: #1489, delete suggesting after deleting old paywall
-                            var products = iapManager.suggestedProducts()
-                            products.insert(.Features.appleTV)
-                            return products
-                        }()
+                        reason: $paywallReason
                     )
                 }
             }
@@ -91,12 +86,7 @@ private extension ProfileStorageSection {
                     Text(Strings.Modules.General.Rows.appletv_compound)
                     PurchaseRequiredView(
                         requiring: tvRequirements,
-                        reason: $paywallReason,
-                        suggesting: { // TODO: #1489, delete suggesting after deleting old paywall
-                            var products = iapManager.suggestedProducts(filter: .onlyComplete)
-                            products.insert(.Features.appleTV)
-                            return products
-                        }()
+                        reason: $paywallReason
                     )
                 }
             }
