+++ app-apple/Package/Sources/CommonLibrary/Extensions/ConfigManager+Extensions.swift
@@ -4,10 +4,10 @@
 
 extension ConfigManager {
     public var canImportToTV: Bool {
-        isActive(.newPaywall) && isActive(.tvWebImport)
+        isActive(.tvWebImport)
     }
 
     public var canSendToTV: Bool {
-        isActive(.newPaywall) && isActive(.tvSendTo)
+        isActive(.tvSendTo)
     }
 }
