+++ app-apple/Package/Sources/CommonLibrary/IAP/IAPManager+Suggestions.swift
@@ -4,93 +4,134 @@
 
 import CommonUtils
 
-@available(*, deprecated, message: "TODO: #1489, delete old paywall")
 extension IAPManager {
-    public enum SuggestionFilter {
-        case excludingComplete
+    public enum Platform {
+        case iOS
 
-        case includingComplete
+        case macOS
 
-        case onlyComplete
+        case tvOS
     }
 
-    public func suggestedProducts(filter: SuggestionFilter = .includingComplete) -> Set<AppProduct> {
+    public enum SuggestionInclusion {
+        case complete
+
+        case singlePlatformEssentials
+    }
+
+    public func suggestedProducts(
+        for features: Set<AppFeature>,
+        including: Set<SuggestionInclusion> = [.complete, .singlePlatformEssentials]
+    ) -> Set<AppProduct> {
 #if os(iOS)
-        suggestedProducts(for: .iOS, filter: filter)
+        suggestedProducts(for: features, on: .iOS, including: including)
 #elseif os(macOS)
-        suggestedProducts(for: .macOS, filter: filter)
+        suggestedProducts(for: features, on: .macOS, including: including)
 #elseif os(tvOS)
-        fatalError("tvOS: Do not suggest products, paywall unsupported")
+        suggestedProducts(for: features, on: .tvOS, including: including)
 #endif
     }
 }
 
 // for testing
-@available(*, deprecated, message: "TODO: #1489, delete old paywall")
 extension IAPManager {
+
+    // suggest the minimum set of products for the given required features
     func suggestedProducts(
-        for platform: Platform,
-        filter: SuggestionFilter,
+        for features: Set<AppFeature>,
+        on platform: Platform,
+        including: Set<SuggestionInclusion>,
         asserting: Bool = false
     ) -> Set<AppProduct> {
-        guard !purchasedProducts.contains(.Essentials.iOS_macOS) else {
-            if asserting {
-                assertionFailure("Suggesting 'Essentials' to former all platforms purchaser?")
-            }
-            return []
-        }
         guard !purchasedProducts.contains(where: \.isComplete) else {
             if asserting {
-                assertionFailure("Suggesting 'Essentials' to complete version purchaser?")
+                assertionFailure("Suggesting products to complete version purchaser?")
             }
             return []
         }
 
         var suggested: Set<AppProduct> = []
 
-        if filter != .onlyComplete {
+        // prioritize eligible features from non-essential products
+        let nonEssentialProducts = features.flatMap(\.nonEssentialProducts)
+        suggested.formUnion(nonEssentialProducts)
+        let nonEssentialEligibleFeatures = Set(nonEssentialProducts.flatMap(\.features))
+
+        //
+        // suggest essential packages if:
+        //
+        // - never purchased any
+        // - non-essential eligible features don't include required essential features
+        //
+        let essentialFeatures = features.filter(\.isEssential)
+        if !didPurchaseEssentials(on: platform) &&
+            !nonEssentialEligibleFeatures.isSuperset(of: essentialFeatures) {
             switch platform {
             case .iOS:
-                guard !purchasedProducts.contains(.Essentials.iOS) else {
-                    if asserting {
-                        assertionFailure("Suggesting 'Essentials iOS' to former iOS purchaser?")
-                    }
-                    return []
-                }
+                // suggest both platforms if never purchased
                 if !purchasedProducts.contains(.Essentials.macOS) {
                     suggested.insert(.Essentials.iOS_macOS)
                 }
-                suggested.insert(.Essentials.iOS)
-            case .macOS:
-                guard !purchasedProducts.contains(.Essentials.macOS) else {
-                    if asserting {
-                        assertionFailure("Suggesting 'Essentials macOS' to former macOS purchaser?")
-                    }
-                    return []
+                // suggest iOS to former macOS purchasers
+                let suggestsSinglePlatform = including.contains(.singlePlatformEssentials) || purchasedProducts.contains(.Essentials.macOS)
+                if suggestsSinglePlatform && !purchasedProducts.contains(.Essentials.iOS) {
+                    suggested.insert(.Essentials.iOS)
                 }
+            case .macOS:
+                // suggest both platforms if never purchased
                 if !purchasedProducts.contains(.Essentials.iOS) {
                     suggested.insert(.Essentials.iOS_macOS)
                 }
-                suggested.insert(.Essentials.macOS)
+                // suggest macOS to former iOS purchasers
+                let suggestsSinglePlatform = including.contains(.singlePlatformEssentials) || purchasedProducts.contains(.Essentials.iOS)
+                if suggestsSinglePlatform && !purchasedProducts.contains(.Essentials.macOS) {
+                    suggested.insert(.Essentials.macOS)
+                }
             case .tvOS:
-                fatalError("Do not present paywall on tvOS")
+                // suggest both platforms if never purchased
+                if !purchasedProducts.contains(where: \.isEssentials) {
+                    suggested.insert(.Essentials.iOS_macOS)
+                }
             }
         }
 
-        if filter != .excludingComplete && purchasedProducts.isEligibleForComplete {
+        let suggestsComplete: Bool
+        switch platform {
+        case .tvOS:
+            //
+            // "essential" features are not accessible from the
+            // TV, therefore selling the "complete" packages is misleading
+            // for TV-only customers. only offer them if some "essential"
+            // feature is required, because it means that the iOS/macOS app
+            // is also installed
+            //
+            // TODO: #103/partout, set always true because all features will be accessible on TV by importing a .json created elsewhere
+            suggestsComplete = !essentialFeatures.isEmpty
+        default:
+            suggestsComplete = true
+        }
+
+        // suggest complete packages if eligible
+        if including.contains(.complete) && suggestsComplete && isEligibleForComplete {
             suggested.insert(.Complete.Recurring.yearly)
             suggested.insert(.Complete.Recurring.monthly)
             suggested.insert(.Complete.OneTime.lifetime)
         }
 
+        // strip purchased (paranoid check)
+        suggested.subtract(purchasedProducts)
+
         return suggested
     }
-}
 
-private extension Collection where Element == AppProduct {
-    var isEligibleForComplete: Bool {
-        !contains {
-            $0.isComplete || $0.isEssentials || $0 == .Features.appleTV
+    func didPurchaseEssentials(on platform: Platform) -> Bool {
+        switch platform {
+        case .iOS:
+            return purchasedProducts.contains(.Essentials.iOS) || purchasedProducts.contains(.Essentials.iOS_macOS)
+        case .macOS:
+            return purchasedProducts.contains(.Essentials.macOS) || purchasedProducts.contains(.Essentials.iOS_macOS)
+        case .tvOS:
+            return purchasedProducts.contains(where: \.isEssentials)
         }
     }
 }
